% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneModel.R
\name{plotGeneModel}
\alias{plotGeneModel}
\title{Gene and mutation plotting}
\usage{
plotGeneModel(gff, snpEffVCF, geneID, upstreamBuffer = 1000,
  downstreamBuffer = 500, features2plot = c("exon", "five_prime_UTR",
  "three_prime_UTR"), colors = c("steelblue3", "lightsteelblue1",
  "lightsteelblue1"), mutations2annotate = c("missense", "stop", "start"),
  pchMutation = c(2, 8, 8), colMutation = c("darkblue", "darkred", "green"),
  orientation = NULL, windowSize = 100, stepSize = 10, scaleBar = 0.1,
  ...)
}
\arguments{
\item{gff}{The gff file, must contain the following column names:
chr, type, start, end, orientation, info}

\item{snpEffVCF}{The snpeff annotated vcf file, must contain the following column names:
CHR, POS, INFO, where info is the snpeff annotations}

\item{upstreamBuffer}{How many bp upstream should be plotted}

\item{downstreamBuffer}{How many bp downstream should be plotted}

\item{features2plot}{What gff elements should be plotted?}

\item{colors}{Colors for exon, 3'UTR and 5'UTR respectively}

\item{mutations2annotate}{Partial string matching for types of differences to annotate}

\item{pchMutation}{Shape of points that specify annotated sites}

\item{colMutation}{Color of points that specify annotated sites}

\item{windowSize}{Sliding window (for difference densities) window size}

\item{stepSize}{Sliding window step size}

\item{scaleBar}{If NULL, do not plot a scale bar, otherwise, specify the proportion of
total gene space that the scale bar covers}

\item{...}{additional arguments, not currently in use.}
}
\value{
a dataframe containing the annotated vcf file of Hal2 Fil2 differences
for the gene specified in geneID
}
\description{
\code{plotGeneModel} Given a gff file and snpeff annotated snp file,
plot the mutations pertaining to a specific gene.
}
\examples{
plotGeneModel(gff = gf, snpEffVCF = sev, geneID = "Pahal.C00786", windowSize=200, stepSize = 5)
}

